Created /etc/init.d file to start the server.

+++ server/temperpi.server.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 # Starts and stops temperpi.server 
 
-TEMPERPI_SERVER_PATH=/Users/rui/git/temperpi/server
+TEMPERPI_SERVER_PATH=/home/ruibm/git/temperpi/server
 MAIN=main.py
 PIDFILE=$MAIN.pid
 
@@ -12,17 +12,21 @@ case "$1" in
     if [ -f $PIDFILE ] ; then
       $0 stop
     fi
-    python $MAIN &> /dev/null &
+    python $MAIN > /dev/null 2>&1 & 
     PID=$!
     echo "$PID" > $PIDFILE
     echo "Successfully started running [$0] with pid [$PID]."
     ;;
 
   stop)
-    PID=`cat $PIDFILE`
-    kill -KILL $PID
-    rm $PIDFILE
-    echo "Successfully killed [$0] with pid [$PID]."
+    if [ -f $PIDFILE ] ; then
+      PID=`cat $PIDFILE`
+      kill -KILL $PID
+      rm $PIDFILE
+      echo "Successfully killed [$0] with pid [$PID]."
+    else
+      $0 status
+    fi
 
     ;;
 
@@ -38,7 +42,7 @@ case "$1" in
       ps -p $PID
       exit 0
     else
-      echo "[$0] is not running"
+      echo "[$0] is not running."
       exit 1
     fi
     ;;
